api: Add authentication middleware to post/comment api

+++ api/src/api/post/post.controllers.ts
@@ -2,14 +2,22 @@ import { createPost, getAllPosts, updatePost, deletePost, getPost } from "api/po
 import HttpError from "utils/http-error";
 
 import type { Request, Response, NextFunction } from "express";
+import type { RequestWithToken } from "middlewares/auth";
 
-const CreatePost = async (req: Request, res: Response, next: NextFunction) => {
+const CreatePost = async (req: RequestWithToken, res: Response, next: NextFunction) => {
   try {
-    const { title, content, authorId, url } = req.body;
-    if (!title || !content || !authorId) {
+    const { title, content, postPhotos } = req.body;
+    const userId = req.payload?.id || "";
+
+    if (!userId) {
+      throw new HttpError("Unauthorized", 401);
+    }
+
+    if (!title || !content) {
       throw new HttpError("Missing required fields", 400);
     }
-    const newPost = await createPost(authorId, title, content, url ? url : url);
+
+    const newPost = await createPost(userId, title, content, postPhotos);
     res.status(201).json({
       newPost,
     });
@@ -17,6 +25,7 @@ const CreatePost = async (req: Request, res: Response, next: NextFunction) => {
     next(error);
   }
 };
+
 const GetPosts = async (req: Request, res: Response, next: NextFunction) => {
   try {
     const posts = await getAllPosts();
@@ -27,6 +36,7 @@ const GetPosts = async (req: Request, res: Response, next: NextFunction) => {
     next(error);
   }
 };
+
 const GetPost = async (req: Request, res: Response, next: NextFunction) => {
   try {
     const { id } = req.params;
@@ -41,17 +51,28 @@ const GetPost = async (req: Request, res: Response, next: NextFunction) => {
     next(error);
   }
 };
-const UpdatePost = async (req: Request, res: Response, next: NextFunction) => {
+
+const UpdatePost = async (req: RequestWithToken, res: Response, next: NextFunction) => {
   try {
     const { id } = req.params;
-    const { title, content, url } = req.body;
+    const userId = req.payload?.id || "";
+    const { title, content, postPhotos } = req.body;
+
     if (!title || !content) {
       throw new HttpError("Missing required fields", 400);
     }
-    const upPost = await updatePost(id, title, content, url ? url : url);
-    if (!upPost) {
+
+    const post = await getPost(id);
+    if (!post) {
       throw new HttpError(`Post with id = ${id} does not exist`, 404);
     }
+
+    if (post.authorId !== userId) {
+      throw new HttpError("Unauthorized", 401);
+    }
+
+    const upPost = await updatePost(id, title, content, postPhotos);
+
     res.status(201).json({
       upPost,
     });
@@ -60,15 +81,23 @@ const UpdatePost = async (req: Request, res: Response, next: NextFunction) => {
   }
 };
 
-const DeletePost = async (req: Request, res: Response, next: NextFunction) => {
+const DeletePost = async (req: RequestWithToken, res: Response, next: NextFunction) => {
   try {
     const { id } = req.params;
-    const deletedPost = await deletePost(id);
-    if (!deletePost) {
+    const userId = req.payload?.id || "";
+
+    const post = await getPost(id);
+    if (!post) {
       throw new HttpError(`Post with id = ${id} does not exist`, 404);
     }
+
+    if (post.authorId !== userId) {
+      throw new HttpError("Unauthorized", 401);
+    }
+
+    await deletePost(id);
     res.status(201).json({
-      deletedPost,
+      message: "Post deleted",
     });
   } catch (error) {
     next(error);
